workspace/requirements.txt
spotipy
google-cloud-storage
python-dotenv